+++ CHANGELOG.rst
@@ -1,5 +1,10 @@
 LATEST CHANGES
 ==============
+
+2024-04-19
+----------
+- GNSS Fusion temporarely disabled
+
 2024-04-11
 ----------
 - Add parameter `pos_tracking.reset_odom_with_loop_closure` to automatically reset odometry when a lood closure is detected